Pull YouTube functionality into its own .js file, which is loaded dynamically only when on a YouTube page.
Now finds and puts RSS feeds for the user's/channel's subscriptions and favourites in the unsorted section of the menu, and  feeds for their playlists in the sorted section. The sorted section now has a menu header (given to us when we requested the feeds via the YouTube API) as well as  the separator.